% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coorClimR.R
\name{convertNeotomaSDToDF}
\alias{convertNeotomaSDToDF}
\title{Get Neotoma Occurrence Data
Function uses the Neotoma API SampleData endpoint}
\usage{
convertNeotomaSDToDF(taxonname, ageold = "", ageyoung = "", loc = "",
  gpid = "", altmin = "", altmax = "")
}
\arguments{
\item{taxonname}{String The name of the taxonomic grouping that you wish tok query neotoma for.  Matches a taxon in the neotoma database}

\item{ageold}{Integer Oldest age, as calendar years before present, to include in results from neotoma.}

\item{ageyoung}{Integer Youngest age, as calendar years before present, to include in results.}

\item{loc}{A list of the form longitudeWest, latitudeSouth, longitudeEast, latitudeNorth that represents a bounding box in which to search for occurrences in Neotoma}

\item{gpid}{Integer Limit occurrences to a geopolitical entity id. Valid values provided by GeoPoliticalUnits database table in Neotoma.}

\item{altmin}{Integer Minimum site altitude in meters.}

\item{altmax}{Integer Maximum site altitude in meters.}
}
\value{
data.frame.  Columns: sampleID (actually a Neotoma dataset ID), Latitude, Longitude, Age
}
\description{
Get Neotoma Occurrence Data
Function uses the Neotoma API SampleData endpoint
}
\examples{
convertNeotomaSDToDF("bison bison")
convertNeotomaSDToDF("picea")
}

