% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coorClimR.R
\name{convertVertnettoDF}
\alias{convertVertnettoDF}
\title{Use Vertnet API to get data, and return only specific columns needed.}
\usage{
convertVertnettoDF(taxonname, genus = "", species = "", state = "",
  limit = "")
}
\arguments{
\item{taxonname}{string: Name(s) of the taxonomic grouping that you wish to query Vertnet for.}

\item{genus}{string: Target genus name(s).}

\item{species}{string: Target species name(s).}

\item{state}{string: Target state name(s).}

\item{limit}{numeric: Number of results that you would like to accept. Defaults to 10000 if left empty.}
}
\value{
output data.frame: Lat, Lon, and Age data.
}
\description{
Use Vertnet API to get data, and return only specific columns needed.
}
\examples{
convertVertnettoDF("bison")
convertVertnettoDF("(kansas state OR KSU)", limit = 200)
convertVertnettoDF(genus = "mustela", species = "(nivalis OR erminea)")

}

