% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coorClimR.R
\name{getDataRow}
\alias{getDataRow}
\title{Get climate data for a single space-time coordnate}
\usage{
getDataRow(x, y, t, producer = "", model = "", modelVersion = "",
  variableType = "", variableUnits = "", variablePeriod = "",
  variablePeriodType = "", averagingPeriod = "", averagingPeriodType = "",
  resolution = "", siteID = "", siteName = "", sampleID = "",
  verbose = F)
}
\arguments{
\item{x}{A double precision longitude value.}

\item{t}{A double precision or integer number representing years before 1950, can be negative to represent time since 1950
API filter parameters}

\item{model}{A string specifying the model from which the climate data was created.  Default = "" (all)}

\item{modelVersion}{A string representing the model version that produced the output. Default = "" (all)}

\item{variableType}{A string representing the type of variable to return. Default = "" (all)}

\item{variableUnits}{A string representing the units in which the variables are measures. Default = "" (all)}

\item{variablePeriod}{An integer representing the measuring period for the variable. Default = "" (all)}

\item{variablePeriodType}{A string representing the type of measuring period for the variable. Default = "" (all)}

\item{averagingPeriodType}{A string representing the type of period over which the data has been averaged.  Default = "" (all)}

\item{resolution}{A double precision value representing the native resolution at which the climate data was produced and stored.  Default = "" (all)}

\item{siteID}{A string or integer identifier for the site at the x-y location of the sample.  Default = "" (none)}

\item{siteName}{A string or integer identifier for the site at the x-y location of the sample. Default = "" (none)}

\item{sampleID}{A string or integer identifier for the sample at the space-time location.  Default = "" (none)}

\item{averaingPeriod}{An integer representing the period of which the data has been averaged. Default= "" (all)}
}
\value{
Outputs a data.frame representation of the api response.  Columns:
 From Database: variableunits, variablePeriodType, VariableType, variableID, Producer, sourceID, ModelVersion, tableName, VariableDescription, averagingPeriod, averagingPeriodType, Model, tableName
 Optional Identifiers:  siteName, sampleID, siteID
 Response Values: value, latitude, longitude, yearsBP
}
\description{
Get climate data for a single space-time coordnate
}
\examples{
getData(-122, 37, 16000, siteName="mySite", variableType="Maximum Temperature")
getData(-90, 40, 900, siteName="myOtherSite", sampleID=1291, variableType="Precipitation")
}

