% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coorClimR.R
\name{makeTSPlot}
\alias{makeTSPlot}
\title{Plot a climate variable through time}
\usage{
makeTSPlot(climateDF, responseVariable = "Precipitation",
  responsePeriod = 1, title = paste(responseVariable, responsePeriod,
  "vs. Time"), plotAVG = T, plotSD = TRUE, pointColor = "gray",
  lineColor = "red")
}
\arguments{
\item{climateDF}{A data.frame object produced by getData that has the climate data you wish to plot}

\item{responseVariable}{The name of the variable for the Y axis of the plot}

\item{responsePeriod}{The measurement period of the variable for the Y axis of the plot}

\item{title}{The title you want to put on the plot}

\item{plotAVG}{Boolean flag for plotting the mean of each yearsBP bin (binned average)}

\item{plotSD}{Boolean flag for plotting the standard deviation of year yearsBP bin(binned SD)}

\item{pointColor}{Color string representing the color in which to plot the individual points}

\item{lineColor}{If plotAVG and/or plotSD are true, then those lines will be plotted in this color}
}
\value{
data.frame with the binned averages, std, and median for the dataset
}
\description{
Plot a climate variable through time
}
\examples{
 sequoia <- queryNeotoma("sequoia")
 makeTSPlot(sequoia, lineColor="green")
}

