% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coorClimR.R
\name{queryAll}
\alias{queryAll}
\title{Query multiple databases at a time. Merge results.}
\usage{
queryAll(taxonname)
}
\arguments{
\item{taxonname}{string: Name(s) of the taxonomic grouping that you wish to query Vertnet for.}
}
\value{
output data.frame: Combined results from all databases
}
\description{
Query multiple databases at a time. Merge results.
}
\examples{
queryAll("bison bison")
}

