% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coorClimR.R
\name{queryNeotoma}
\alias{queryNeotoma}
\title{Get Climate Data for Neotoma Occurrences}
\usage{
queryNeotoma(taxonname, ageold = "", ageyoung = "", loc = "", gpid = "",
  altmin = "", altmax = "", producer = "", model = "",
  modelVersion = "", variableType = "", variableUnits = "",
  variablePeriod = "", variablePeriodType = "", averagingPeriod = "",
  averagingPeriodType = "", resolution = "")
}
\arguments{
\item{taxonname}{String The name of the taxonomic grouping that you wish tok query neotoma for.  Matches a taxon in the neotoma database}

\item{ageold}{Integer Oldest age, as calendar years before present, to include in results from neotoma.}

\item{ageyoung}{Integer Youngest age, as calendar years before present, to include in results.}

\item{loc}{A list of the form longitudeWest, latitudeSouth, longitudeEast, latitudeNorth that represents a bounding box in which to search for occurrences in Neotoma}

\item{gpid}{Integer Limit occurrences to a geopolitical entity id. Valid values provided by GeoPoliticalUnits database table in Neotoma.}

\item{altmin}{Integer Minimum site altitude in meters.}

\item{altmax}{Integer Maximum site altitude in meters.}

\item{model}{A string specifying the model from which the climate data was created.  Default = "" (all)}

\item{modelVersion}{A string representing the model version that produced the output. Default = "" (all)}

\item{variableType}{A string representing the type of variable to return. Default = "" (all)}

\item{variableUnits}{A string representing the units in which the variables are measures. Default = "" (all)}

\item{variablePeriod}{An integer representing the measuring period for the variable. Default = "" (all)}

\item{variablePeriodType}{A string representing the type of measuring period for the variable. Default = "" (all)}

\item{averagingPeriodType}{A string representing the type of period over which the data has been averaged.  Default = "" (all)}

\item{resolution}{A double precision value representing the native resolution at which the climate data was produced and stored.  Default = "" (all)}

\item{averaingPeriod}{An integer representing the period of which the data has been averaged. Default= "" (all)}
}
\value{
Outputs a data.frame representation of the api response.  Columns:
 From Database: variableunits, variablePeriodType, VariableType, variableID, Producer, sourceID, ModelVersion, tableName, VariableDescription, averagingPeriod, averagingPeriodType, Model, tableName
 Optional Identifiers:  siteName, sampleID, siteID
 Response Values: value, latitude, longitude, yearsBP
}
\description{
Get Climate Data for Neotoma Occurrences
}
\examples{
queryNeotoma("bison bison")
queryNeotoma("sedum", altmax=1500, variablePeriod=1, resolution=0.5)
}

