% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coorClimR.R
\name{queryVertnet}
\alias{queryVertnet}
\title{Get climate data for Vertnet occurrences.}
\usage{
queryVertnet(taxonname, genus = "", species = "", state = "",
  limit = "")
}
\arguments{
\item{taxonname}{string: Name(s) of the taxonomic grouping that you wish to query Vertnet for.}

\item{genus}{string: Target genus name(s).}

\item{species}{string: Target species name(s).}

\item{state}{string: Target state name(s).}

\item{limit}{numeric: Number of results that you would like to accept. Defaults to 10000 if left empty.}
}
\value{
output data.frame: The data which you seek.
}
\description{
Get climate data for Vertnet occurrences.
}
\examples{
queryVertnet("bison")
queryVertnet("(kansas state OR KSU)", limit = 200)
queryVertnet(genus = "mustela", species = "(nivalis OR erminea)")

}

